+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 Here's the change logs of `regex-go`.
 
+## v0.4.0
+
+- Breaking change: default version now changed to umd format, if you want to use esm version, please use `regex-go/esm`.
+
+## v0.3.1
+
+- Feat: add ip loose testers (fuzzy matching).
+
+## v0.3.0
+
+- Feat: add ip related testers.
+
+- Feat: export all testers.
+
 ## 0.2.0
 
 - Feat: add `containsChinese` method.